+++ src/ReaderTaskEither.ts
@@ -3,7 +3,7 @@
  */
 import { Alt3, Alt3C } from './Alt'
 import { Applicative3, Applicative3C } from './Applicative'
-import { Apply3 } from './Apply'
+import { apComposition, Apply3 } from './Apply'
 import { Bifunctor3 } from './Bifunctor'
 import * as E from './Either'
 import { identity, pipe, Predicate, Refinement } from './function'
@@ -24,7 +24,6 @@ import { Semigroup } from './Semigroup'
 import { Task } from './Task'
 import * as TE from './TaskEither'
 import * as ValidationT from './ValidationT'
-import * as EitherT from './EitherT'
 
 import Reader = R.Reader
 import Either = E.Either
@@ -265,7 +264,7 @@ export function getReaderTaskValidationApplicative<E>(S: Semigroup<E>): Applicat
     URI,
     _E: undefined as any,
     map,
-    ap: EitherT.ap(RT.applyReaderTask, E.getValidationApplicative(S)),
+    ap: apComposition(RT.applyReaderTask, E.getValidationApplicative(S)),
     of
   }
 }