+++ src/index.ts
@@ -24,7 +24,6 @@ import * as contravariant from './Contravariant'
 import * as date from './Date'
 import * as distributiveLattice from './DistributiveLattice'
 import * as either from './Either'
-import * as eitherT from './EitherT'
 import * as eq from './Eq'
 import * as extend from './Extend'
 import * as field from './Field'
@@ -181,10 +180,6 @@ export {
    * @since 2.0.0
    */
   either,
-  /**
-   * @since 2.0.0
-   */
-  eitherT,
   /**
    * @since 2.0.0
    */